chore: 감정분석 결과 정수로 반환하기

+++ backend/src/main/java/com/starbooks/backend/emotion/service/EmotionService.java
@@ -78,8 +78,8 @@ public class EmotionService {
             return new EmotionPoint(0.0, 0.0);
         }
 
-        double weightedValence = clamp(sumValence / sumWeight, -5.0, 5.0);
-        double weightedArousal = clamp(sumArousal / sumWeight, -5.0, 5.0);
+        int weightedValence = (int)clamp(sumValence / sumWeight, -5.0, 5.0);
+        int weightedArousal = (int)clamp(sumArousal / sumWeight, -5.0, 5.0);
 
         return new EmotionPoint(weightedValence, weightedArousal);
     }
